Just adding the select, to select a period :P
