+++ components/home/HomeAnalysisOptions.vue
@@ -2,15 +2,6 @@
     import { useAnalysisOptions, ReportType } from "@/stores/analysisOptions";
 
     const analysisStore = useAnalysisOptions();
-
-    function handleReportTypeChange(payload: string | number) {
-        const value = payload.toString();
-
-        if (value !== ReportType.SABER11 && value !== ReportType.SABERPRO)
-            return;
-
-        analysisStore.setReportType(value);
-    }
 </script>
 
 <template>
@@ -26,27 +17,10 @@
                         meaninful insights.</span
                     >
                 </div>
-                <div>
-                    <span class="font-semibold">Report Type</span>
-                    <Select
-                        :default-value="ReportType.SABER11"
-                        @update:model-value="handleReportTypeChange"
-                    >
-                        <SelectTrigger class="mt-2">
-                            <SelectValue placeholder="Select the Report Type" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            <SelectItem
-                                v-for="type in ReportType"
-                                :key="type"
-                                :value="type"
-                            >
-                                {{ type }}
-                            </SelectItem>
-                        </SelectContent>
-                    </Select>
+                <div class="flex flex-col gap-4">
+                    <HomeSelectReportType />
+                    <HomeSelectPeriod />
                 </div>
-                <HomeSelectPeriod />
             </div>
             <div class="flex flex-col gap-4">
                 <HomeSelectDepartment />
